pyproject.toml
[tool.poetry]
name = "qa-guru-mobile-browserstack"
version = "0.1.0"
description = "mobile end-to-end testing on browserstack with selene + appium + pytest"
authors = ["Rustam Dzhafarov <rustamd74@gmail.com>"]
readme = "README.md"
packages = [{include = "qa_guru_mobile_browserstack"}]

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.5.0"
selene = {version = "^2.0.0-beta.17", allow-prereleases = true}
Appium-Python-Client = "^2.8.1"
pytest = "^7.1.3"
pytest-xdist = "^3.2.1"
allure-pytest = "^2.10.0"
requests = "^2.28.1"
pydantic = "^1.10.5"
python-dotenv = "^1.0.0"


[tool.poetry.dev-dependencies]
black = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
skip-string-normalization = true


[tool.pytest.ini_options]
addopts = " --alluredir=allure-results "
